+++ Assets/HotAssets/Scripts/GamePlay/Logic/Common/GamePlayEvent.cs
@@ -5,25 +5,25 @@ namespace HotAssets.Scripts.GamePlay.Logic.Common
 {
     public class GamePlayEvent
     {
-        public static readonly Int16 ECreateRoleAll = ++GameEvent.UIEventId;
-        public static readonly Int16 ERenderAllRole = ++GameEvent.UIEventId;
-        public static readonly Int16 ECreateMapRender = ++GameEvent.UIEventId;
-        public static readonly Int16 EFightLoadingFinish = ++GameEvent.UIEventId;
-        public static readonly Int16 ECameraFocus = ++GameEvent.UIEventId;
-        public static readonly Int16 ERenderBullet = ++GameEvent.UIEventId;
-        public static readonly Int16 EStopRenderBullet = ++GameEvent.UIEventId;
-        public static readonly Int16 ERenderMonster = ++GameEvent.UIEventId;
-        public static readonly Int16 EStopRenderRole = ++GameEvent.UIEventId;
-        public static readonly Int16 EPlayEffect = ++GameEvent.UIEventId;
-        public static readonly Int16 ERenderAoe = ++GameEvent.UIEventId;
-        public static readonly Int16 EStopRenderAoe = ++GameEvent.UIEventId;
+        public static readonly int ECreateRoleAll = ++GameEvent.UIEventId;
+        public static readonly int ERenderAllRole = ++GameEvent.UIEventId;
+        public static readonly int ECreateMapRender = ++GameEvent.UIEventId;
+        public static readonly int EFightLoadingFinish = ++GameEvent.UIEventId;
+        public static readonly int ECameraFocus = ++GameEvent.UIEventId;
+        public static readonly int ERenderBullet = ++GameEvent.UIEventId;
+        public static readonly int EStopRenderBullet = ++GameEvent.UIEventId;
+        public static readonly int ERenderMonster = ++GameEvent.UIEventId;
+        public static readonly int EStopRenderRole = ++GameEvent.UIEventId;
+        public static readonly int EPlayEffect = ++GameEvent.UIEventId;
+        public static readonly int ERenderAoe = ++GameEvent.UIEventId;
+        public static readonly int EStopRenderAoe = ++GameEvent.UIEventId;
         
-        public static readonly Int16 EUpdateFrame = ++GameEvent.UIEventId;
-        public static readonly Int16 ESvrRoomPlayerInfoUpdate = ++GameEvent.UIEventId;
+        public static readonly int EUpdateFrame = ++GameEvent.UIEventId;
+        public static readonly int ESvrRoomPlayerInfoUpdate = ++GameEvent.UIEventId;
         
-        public static readonly Int16 ELoadSceneSuccess = ++GameEvent.UIEventId;
-        public static readonly Int16 ELoadSceneFailure = ++GameEvent.UIEventId;
-        public static readonly Int16 ELoadSceneUpdate = ++GameEvent.UIEventId;
-        public static readonly Int16 EGameOver = ++GameEvent.UIEventId;
+        public static readonly int ELoadSceneSuccess = ++GameEvent.UIEventId;
+        public static readonly int ELoadSceneFailure = ++GameEvent.UIEventId;
+        public static readonly int ELoadSceneUpdate = ++GameEvent.UIEventId;
+        public static readonly int EGameOver = ++GameEvent.UIEventId;
     }
 }
